+++ src/components/MessageNode/Fluorite5Node.js
@@ -2,7 +2,8 @@ import React, { PropTypes } from 'react';
 import IPropTypes from 'react-immutable-proptypes';
 import IconButton from 'material-ui/IconButton';
 import Info from 'material-ui/svg-icons/action/info-outline';
-import { grey300 } from 'material-ui/styles/colors';
+import { grey300, grey800 } from 'material-ui/styles/colors';
+import Dice6 from '../icons/Dice6';
 
 const Style = {
     Container: {
@@ -18,16 +19,75 @@ const Style = {
         width: 16,
         height: 16,
     },
+    DiceResults: {
+        marginRight: '8',
+    },
+    DiceTooltip: {
+        display: 'flex',
+        backgroundColor: 'white',
+        color: 'black',
+    },
+    Tooltip: {
+        padding: 0,
+    },
 };
 
-const diceToTooltip = (dice) =>
-    dice
-        .map((result) => {
-            const faces = result.get('faces');
-            const results = result.get('results');
-            return `${results.size}d${faces}=[${results.join(',')}]`;
-        })
-        .join(' ');
+const DiceResult = (props) => {
+    const {
+        faces,
+        result,
+    } = props;
+
+    if (faces === 6) {
+        return <Dice6 color={grey800} n={result} />;
+    }
+
+    return null;
+};
+DiceResult.propTypes = {
+    faces: PropTypes.number.isRequired,
+    result: PropTypes.number.isRequired,
+};
+
+const DiceResults = (props) => {
+    const {
+        faces,
+        results,
+    } = props;
+
+    if (faces === 6) {
+        return (
+            <div style={Style.DiceResults}>
+                {results.map((result, i) => <DiceResult key={i} faces={faces} result={result} />)}
+            </div>
+        );
+    }
+
+    return <div>{`${results.size}d${faces}=[${results.join(',')}]`}</div>;
+};
+DiceResults.propTypes = {
+    faces: PropTypes.number.isRequired,
+    results: IPropTypes.listOf(PropTypes.number).isRequired,
+};
+
+const DiceTooltip = (props) => {
+    const {
+        dice,
+    } = props;
+
+    return (
+        <div style={Style.DiceTooltip}>
+            {dice.map((result, i) => {
+                const faces = result.get('faces');
+                const results = result.get('results');
+                return <DiceResults key={i} faces={faces} results={results} />;
+            })}
+        </div>
+    );
+};
+DiceTooltip.propTypes = {
+    dice: IPropTypes.list,
+};
 
 const Fluorite5Node = (props) => {
     const {
@@ -39,8 +99,9 @@ const Fluorite5Node = (props) => {
         <IconButton
             iconStyle={Style.Icon}
             style={Style.IconButton}
-            tooltip={diceToTooltip(dice)}
+            tooltip={<DiceTooltip dice={dice} />}
             tooltipPosition="top-center"
+            tooltipStyles={Style.Tooltip}
         >
             <Info />
         </IconButton>